refactor: move instance status policy to models
