+++ crates/dcc-mcp-transport/Cargo.toml
@@ -25,6 +25,7 @@ tokio = { workspace = true }
 rmp-serde = { workspace = true }
 mdns-sd = { workspace = true, optional = true }
 dcc-mcp-pybridge = { workspace = true, optional = true }
+dcc-mcp-models = { workspace = true }
 sysinfo.workspace = true
 ipckit.workspace = true
 workspace-hack = { version = "0.1", path = "../workspace-hack" }
